Revert "Add Falling Relax with only upper body relaxing (#1527)"
This reverts commit 609a75e2f95740c7478872f1679de27acd137fb0.
