+++ module/strategy/FallRecovery/src/FallRecovery.cpp
@@ -44,7 +44,7 @@ namespace module::strategy {
         on<Provide<FallRecoveryTask>>().then([this] {
             // Plan to relax when falling and get up when on the ground
             // We set the get up priority higher than falling so that relax won't take over while we are getting up
-            emit<Task>(std::make_unique<RelaxWhenFalling>(), 1);
+            // emit<Task>(std::make_unique<RelaxWhenFalling>(), 1); // commented out until it stops causing issues
             emit<Task>(std::make_unique<GetUpWhenFallen>(), 2);
         });
     }
