--- shared/utility/skill/scripts/nugus/RelaxArms.yaml
@@ -1,26 +1,0 @@
-- duration: 25
-  targets:
-    - id: R_SHOULDER_PITCH
-      position: 0
-      gain: .NaN
-      torque: 0
-    - id: L_SHOULDER_PITCH
-      position: 0
-      gain: .NaN
-      torque: 0
-    - id: R_SHOULDER_ROLL
-      position: 0
-      gain: .NaN
-      torque: 0
-    - id: L_SHOULDER_ROLL
-      position: 0
-      gain: .NaN
-      torque: 0
-    - id: R_ELBOW
-      position: 0
-      gain: .NaN
-      torque: 0
-    - id: L_ELBOW
-      position: 0
-      gain: .NaN
-      torque: 0
