+++ shared/utility/skill/scripts/webots/RelaxHead.yaml
@@ -1,10 +1,10 @@
-- duration: 25
-  targets:
-    - id: HEAD_YAW
-      position: 0
-      gain: .NaN
-      torque: 0
-    - id: HEAD_PITCH
-      position: 0
-      gain: .NaN
-      torque: 0
+[
+  {
+    "duration": 25,
+    "targets":
+      [
+        { "id": "HEAD_YAW", "position": 0, "gain": .NaN, "torque": 0 },
+        { "id": "HEAD_PITCH", "position": 0, "gain": .NaN, "torque": 0 },
+      ],
+  },
+]
